Extract tests for the ApiClient
